segundo commit

+++ questao10.py
@@ -2,9 +2,9 @@
 Fazer um algoritmo que efetue o cálculo do valor de uma prestação em atraso, utilizando a fórmula prestação	=
 valor	+	(valor	x	(taxa	:	100)	x	tempo	em	dias)
 '''
-val = int(input("Qual o valor da prestação?"))
+val = float(input("Qual o valor da prestação?"))
 tax = float(input("Qual o valor da taxa?"))
-temp = int(input("Qual o tempo em atraso (dias) ?"))
+temp = float(input("Qual o tempo em atraso (dias) ?"))
 
 prest = val +(val*(tax/100)*temp)
